modelAPI/requirements.txt
functions-framework==3.*
pandas==2.2.2
scikit-learn==1.5.0
joblib==1.4.2
google-cloud-storage==2.16.0